Run TypeScript verification in new context
